+++ frontend/js/gens.ts
@@ -22,7 +22,7 @@ import { API } from "./state/api";
 import { TracksManager } from "./components/tracks_manager/tracks_manager";
 import { InputControls } from "./components/input_controls";
 import { getRenderDataSource } from "./state/data_source";
-import { STYLE } from "./constants";
+import { DEFAULT_VARIANT_THRES as DEFAULT_VARIANT_THRES, STYLE } from "./constants";
 import { SideMenu } from "./components/side_menu/side_menu";
 import {
   SettingsMenu,
@@ -157,6 +157,7 @@ export async function initCanvases({
     api.getChromInfo(),
     api.getChromSizes(),
     startRegion,
+    DEFAULT_VARIANT_THRES,
   );
 
   const renderDataSource = getRenderDataSource(
@@ -185,7 +186,7 @@ export async function initCanvases({
 
   settingsPage.setSources(
     session,
-    () => render({}),
+    render,
     allAnnotSources,
     defaultAnnot,
     () => gensTracks.trackView.getDataTracks(),
@@ -229,6 +230,10 @@ export async function initCanvases({
       session.setColorAnnotation(annotId);
       await gensTracks.trackView.setColorAnnotation(annotId);
       render({});
+    },
+    (rng: Rng) => {
+      gensTracks.setCovYRange(rng);
+      render({dataUpdated: true});
     },
   );